Planned improval for team pace graph lap filtering
Original implementation only uses pick_wo_box. This misses SC and VSC slow laps.

Only using pick_quicklaps is undesirable because it is not robust in mixed conditions races

The best solution is to combine pick_wo_box and pick_track_status.

This is blocked as pick_track_status is susceptible to crashing and needs to be fixed first.