Implement the minimum lap count limit for the compound plot
It is stated in the caveat that only compounds that complete a minimum of one twentieth of eligible laps can be plotted.

However, this is not previously implemented.

This can cause confusion as a selecting a compound may not result in any change on the graph.

It should be noted that this change does not completely eliminate the above behavior. As whether the curve is shown is determined by the number of records for each `TyreLife` value (minimum 3 to be shown).

But this change should eliminate the majority of such cases.